link catalog, product and profile view to routing: -remove unused files -add an container folder -rename somes files -moov some files or folders -regenerate dep. graph
