--- src/components/Footer.tsx
@@ -1,3 +1,0 @@
-export function Footer(children: JSX.Element) {
-  return <>Footer{children}</>;
-}
