--- src/components/Navbar.tsx
@@ -1,3 +1,0 @@
-export function AppMenu() {
-  return;
-}
