+++ src/components/containers/App.test.tsx
@@ -1,6 +1,5 @@
 import { render, screen } from "@testing-library/react";
-import App from "main/App";
-import React from "react";
+import App from "./App";
 
 test("renders learn react link", () => {
   render(<App />);
