+++ src/components/containers/App.tsx
@@ -1,6 +1,6 @@
-import { RootMenu } from "components/Menu/RootMenu";
+import { RootMenu } from "components/menus/RootMenu";
+import { RoutePath } from "main/RoutePath";
 import { Outlet } from "react-router-dom";
-import { RoutePath } from "./RoutePath";
 
 export default function App() {
   return (
