+++ src/components/pages/HomePage.tsx
@@ -3,5 +3,6 @@ import { Page } from "components/pages/Page";
 export const HomePage = () => (
   <Page>
     <h1>Home</h1>
+    <p>Recent offers list</p>
   </Page>
 );
