+++ src/components/views/ProfileView.tsx
@@ -1,0 +1,9 @@
+export function ProfileView() {
+  return (
+    <main>
+      <h1>My Profile (sell or client)</h1>
+
+      <p>...</p>
+    </main>
+  );
+}
