--- src/views/AccountView.tsx
@@ -1,12 +1,0 @@
-export function AccountView() {
-  return (
-    <main>
-      <h1>My Account</h1>
-      <h2>Security</h2>
-      <h2>Invoices</h2>
-      <h2>Offers</h2>
-      <h2>Catalogs</h2>
-      <p>...</p>
-    </main>
-  );
-}
